requirements.txt
absl-py==2.3.1
astunparse==1.6.3
bidict==0.23.1
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dnspython==2.8.0
eventlet==0.40.3
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==12.0.0
pip==23.0.1
protobuf==6.33.0
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-engineio==3.13.2
python-socketio==4.6.1
requests==2.32.5
rich==14.2.0
setuptools==65.5.0
simple-websocket==1.1.0
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
wsproto==1.2.0
